fix: SCF page data + ecosystem resilience + gitcoin2 bronze observe mode
- scf.ts: cast totalPaidUSD SUM to ::numeric before ROUND() — fixes
  'function round(double precision, integer) does not exist' that was
  returning 500 on /api/systems/scf and blanking the SCF page
- ecosystem.ts: wrap gold queries (ecosystem_overview, cross_platform,
  scf_system_profile) in individual .catch() so a stale gold table
  (e.g. gold__cross_platform missing ens_funding) no longer crashes
  the entire endpoint
- gitcoin2.py: add GITCOIN2_OBSERVE_ONLY mode — when set, queries DB
  for existing bronze table row counts and registers them in Dagster's
  event log without re-loading CSVs
- definitions.py: add etl_gitcoin2_full_job (bronze→silver→gold),
  matching the full pipeline jobs for all other platforms

+++ nextjs-dashboard/src/pages/api/systems/scf.ts
@@ -282,7 +282,7 @@ export default async function handler(
         COALESCE(SUM("org.stellar.communityfund.totalAwardedUSD"), 0) as total_awarded_usd,
         COALESCE(SUM("org.stellar.communityfund.totalPaidUSD"), 0) as total_paid_usd,
         ROUND(
-          COALESCE(SUM("org.stellar.communityfund.totalPaidUSD"), 0) / NULLIF(COALESCE(SUM("org.stellar.communityfund.totalAwardedUSD"), 0), 0) * 100,
+          COALESCE(SUM("org.stellar.communityfund.totalPaidUSD"), 0)::numeric / NULLIF(COALESCE(SUM("org.stellar.communityfund.totalAwardedUSD"), 0), 0) * 100,
         1) as payment_rate_pct
       FROM silver_scf_grant_applications
       WHERE "org.stellar.communityfund.awardType" IS NOT NULL AND "org.stellar.communityfund.awardType" != ''
